+++ src/components/layout/contextMenu/ContextMenuFolder.vue
@@ -24,7 +24,7 @@ update-file-name(:pageState="updateFileModalProps" @closeModal="updateFileModalP
 import DeleteFileModal from '@/components/modal/DeleteFileModal.vue';
 import UpdateFileName from '@/components/modal/UpdateFileNameModal.vue';
 import { reactive } from 'vue';
-import { globalMethod } from '@/stores/lin';
+import { globalMethod } from '@/stores/useStore';
 const $globalMethod = globalMethod();
 const $globalF = $globalMethod.$globalFunction;
 const $emitter = $globalMethod.$emitter;
