+++ src/components/modal/MessageModal.vue
@@ -2,7 +2,7 @@
 a-space
 </template>
 <script setup lang="ts">
-import { globalMethod } from '@/stores/lin';
+import { globalMethod } from '@/stores/useStore';
 import { message } from 'ant-design-vue';
 const $globalMethod = globalMethod();
 const $emitter = $globalMethod.$emitter;
